+++ docs/changelog.md
@@ -7,6 +7,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.1.0/), 
 Version numbering adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 - [master]
+  - [Changed] it's now possible to use both ``` and ~~~ for code block marks. As per Markdown rules, the closing mark has to be the same as the opening one.
   - [Added]   first implementation of a progress bar, `-sb` option
   - [Changed] On the command line, commands no more start with '-' or '--' (previous form still taken into account for now)
   - [Fixed]   Closes #7